add test scene, go double

+++ source/AreaLight.h
@@ -21,13 +21,13 @@ public:
     //Color SamplePoint (Vector3d& point, Vector3d& normal) const;
     Vector3d GetNormal() const;
 
-    float Pdf(const IntersectionInfo& info, const Vector3d& out) const;
+    double Pdf(const IntersectionInfo& info, const Vector3d& out) const;
 
     Color NextEventEstimation(const Renderer* renderer, const IntersectionInfo& info) const;
     Color NextEventEstimationMIS(const Renderer* renderer, const IntersectionInfo& info) const;
     Color DirectHitMIS(const Renderer* renderer, const IntersectionInfo& lastInfo, const IntersectionInfo& thisInfo) const;
 
-    Color SampleRay(Ray& ray, Vector3d& Normal, float& areaPdf, float& anglePdf) const;
+    Color SampleRay(Ray& ray, Vector3d& Normal, double& areaPdf, double& anglePdf) const;
     void SamplePoint(Vector3d& point, Vector3d& Normal) const;
 
     void Save(Bytestream& s) const;
@@ -35,7 +35,7 @@ public:
 
     void AddPortal(const Vector3d& pos, const Vector3d& v1, const Vector3d& v2);
 
-    float GetArea() const;
+    double GetArea() const;
     void AddToScene(std::shared_ptr<Scene> scn);
 protected:
     Vector3d pos, c1, c2;
