+++ source/CsgObject.h
@@ -14,7 +14,7 @@ class CsgHit
 public:
     enum HitType { Enter, Exit };
     IntersectionInfo info;
-    float t;
+    double t;
     HitType type;
 };
 
@@ -29,11 +29,11 @@ public:
     virtual BoundingBox GetBoundingBox() const = 0;
     virtual bool GetClippedBoundingBox(const BoundingBox& clipbox, BoundingBox& resultbox) const = 0;
 
-    virtual float Intersect(const Ray& ray) const = 0;
+    virtual double Intersect(const Ray& ray) const = 0;
     virtual bool GenerateIntersectionInfo(const Ray& ray, IntersectionInfo& info) const = 0;
 
     virtual void Translate(const Vector3d& direction) = 0;
-    virtual void Rotate(const Vector3d& axis, float angle) = 0;
+    virtual void Rotate(const Vector3d& axis, double angle) = 0;
 
     virtual std::unique_ptr<CsgObject> Clone() = 0;
 protected:
